+++ packages/charts/src/utils/themes/light_theme.ts
@@ -394,4 +394,13 @@ export const LIGHT_THEME: Theme = {
       },
     },
   },
+  metric: {
+    text: {
+      lightColor: '#E0E5EE',
+      darkColor: '#343741',
+    },
+    barBackground: '#EDF0F5',
+    background: '#FFFFFF',
+    nonFiniteText: 'N/A',
+  },
 };