+++ storybook/stories/wordcloud/1_wordcloud.story.tsx
@@ -270,8 +270,3 @@ export const Example = () => {
     </Chart>
   );
 };
-
-Example.parameters = {
-  background: { disable: true },
-  theme: { disable: true },
-};